Checkout from development to edge branch using -b flag
Ensures we don't need an extra step to pull in recent changes from
development. Without the flag, we end up checking out yesterday's edge
branch copy which will be missing any recent non-composer changes.

With the -b flag, we effectively start a new branch from the time the
job runs and then force push it up at the end. This ensures we gather
all the changes we want.