Patch antibot JS to improve detection of humans
Antibot JS verifies a human by listening for mousemove and touchmove events.  For multistep webforms, it is sometimes possible to click the submit button without moving the mouse causing Antibot JS to block the form submission.  This patch updates event listeners to listen for newer pointerdown (equivalent to mousedown and touchstart) and pointermove (equivalent to mousemove and touchmove) events.  It also makes use of the event.isTrusted property that is set by the browser when it determines that a user triggered the event (trusted) versus a script triggering the event (untrusted).  In this way it may prevent more sophisticated bots (using headless browsers) simulating mouse, touch and keyboard events to bypass antibot protection.