Update PrisonerPaymentsWebformHandler.php
Prevent more than 1 visitor making a payment to a prisoner at the same time.  If there is a pending transaction for a visitor, another visitor who tries to make a payment will be prevented from proceeding and told to try again later.

Ensure prisoner_id exists in prisoner_payment_amount table before checking if prisoner_id has nominated visitor_ids who can make payments to them.  This is needed because a prisoner_id may not exist in the amount table when it is updated by Prism and the prisoner_id is removed.