Update PrisonVisitBookingHandler.php
On the time slot selection step (visit_preferred_day_and_time), make a fresh call to getAvailableSlots() instead of using the ones retrieved earlier that are stored in this->bookingReference.  This is to prevent using possibly stale time slot data.  If no time slots are now available, show a message to the user and prevent further progress in the booking.